+++ src/background.js
@@ -1,8 +1,8 @@
-import {iconListeners} from "./scripts/units/extension-icon/icon.bg";
+import {iconListeners} from "./scripts/units/chrome/extension-icon/icon.bg";
 import {clickActionListener} from "./scripts/process/background/export-manager/exportOnce.bg";
 import {exportAllThreadsListener} from "./scripts/process/background/export-manager/exportAll.bg";
 import {notify} from "./scripts/process/background/notify";
-import {clickCountInit} from "./scripts/units/modals/clickCount.all";
+import {clickCountInit} from "./scripts/units/interface/modals/clickCount.all";
 
 background();
