+++ src/scripts/units/chrome/extension-icon/icon.bg.js
