+++ src/scripts/units/chrome/extension-icon/iconSetter.bg.js
