+++ src/scripts/units/interface/uiEnhancer/phind/historySearch/insertHistoryBar.js
