+++ src/scripts/units/processing/checker/domainChecker.all.js
