+++ src/scripts/units/processing/checker/domainChecker.tab.js
@@ -1,4 +1,4 @@
-import {getHostAndPath} from "../utils/getters.tab";
+import {getHostAndPath} from "../../utils/getters.tab";
 
 export async function domainChecker(domains) {
   const hostAndPath = getHostAndPath();
