+++ src/scripts/units/processing/formatter/formatText.all.js
